Remove usage of SnapKit dependency, add BMError enum type, fix compilation error in example project and fix UI in example project
Change all usage of Snapkit to use iOS provided constraint functionality through anchors. This allows the pod to be much smaller when SnapKit is removed. It also reduces possible issues with dependencies. I also removed usage of SnapKit in the example project and replaced it with anchors so there is no reference to SnapKit any longer and it can be removed as a dependency from this project.

I also fixed the UI for the example code, which had the video being covered by the navigation controller nav bar. I added extra code to take care of the edge case when the VideoPlayViewController is opened in landscape mode so that the nav bar will disappear. I've also added a BMError enum type with a raw value to be printed as a fatal error in case the pod is used in in less than iOS 9.

These changes will cause the pod to no longer work for iOS 8, since I used anchors as the constraint solution. It should be noted that little over 1% of users are still using iOS 8.